requirements.txt
Flask
transformers
sentence-transformers
faiss-cpu
numpy
